+++ packages/runtime/src/parser/parser.ts
@@ -68,6 +68,7 @@ import {
 	EcdsaPkRecover,
 	EcdsaVerify,
 	Ed25519verify,
+	Ed25519verify_bare,
 	EqualTo,
 	Err,
 	Exp,
@@ -136,6 +137,7 @@ import {
 	SetBit,
 	SetByte,
 	Sha256,
+	Sha3_256,
 	Sha512_256,
 	Shl,
 	Shr,
@@ -400,6 +402,8 @@ opCodeMap[7] = {
 	base64_decode: Base64Decode,
 	replace2: Replace2,
 	replace3: Replace3,
+	sha3_256: Sha3_256,
+	ed25519verify_bare: Ed25519verify_bare,
 };
 
 // list of opcodes with exactly one parameter.
@@ -465,6 +469,7 @@ const interpreterReqList = new Set([
 	"sha256",
 	"sha512_256",
 	"keccak256",
+	"sha3_256",
 ]);
 
 const signatureModeOps = new Set(["arg", "args", "arg_0", "arg_1", "arg_2", "arg_3"]);
@@ -509,8 +514,10 @@ const commonModeOps = new Set([
 	"err",
 	"sha256",
 	"keccak256",
+	"sha3_256",
 	"sha512_256",
 	"ed25519verify",
+	"Ed25519verify_bare",
 	"ecdsa_verify",
 	"ecdsa_pk_decompress",
 	"+",