+++ packages/web/src/lib/wallectconnect-mode.ts
@@ -246,11 +246,11 @@ export class WallectConnectSession {
 		}
 		// sign smart signature transaction
 		for (const [index, txn] of txns.entries()) {
-			const singer: Sign = execParams[index];
-			if (singer.sign === SignType.LogicSignature) {
-				singer.lsig.lsig.args = singer.args ? singer.args : [];
+			const signer: Sign = execParams[index];
+			if (signer.sign === SignType.LogicSignature) {
+				signer.lsig.lsig.args = signer.args ? signer.args : [];
 				if (!Array.isArray(signedTxn)) signedTxn = [];
-				signedTxn.splice(index, 0, algosdk.signLogicSigTransaction(txn, singer.lsig).blob);
+				signedTxn.splice(index, 0, algosdk.signLogicSigTransaction(txn, signer.lsig).blob);
 			}
 		}